% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lost_catch.R
\name{estimate_lost_catch}
\alias{estimate_lost_catch}
\title{Estimate Lost Catch
A function to estimate lost catch due to over-fishing
Based on the work of Cheung & Srinivasan 
The data.frame must also include}
\usage{
estimate_lost_catch(data, msy, max_catch)
}
\arguments{
\item{data}{A data.frame with two columns: 1. a catch time series normalized to the max catch of the time series. 
That is, each year of landings expressed as a % of the maximum annual catch. 2. A column that categorizes each year of catch as
either developing, exploited, over-exploited, collapsed, or rebuilding.}
}
\description{
Estimate Lost Catch
A function to estimate lost catch due to over-fishing
Based on the work of Cheung & Srinivasan 
The data.frame must also include
}
\keyword{MSY,}
\keyword{Maximum}
\keyword{Sustainable}
\keyword{Yield,}
\keyword{catch}
\keyword{lost}
