% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimate_msy}
\alias{estimate_msy}
\title{Estimate MSY
A function to estimate Maximum Sustainable Yield from a time series of landings
The function requires only the maximum catch of a time-series.
Based on the work of Cheung & Srinivasan who find a statistical relationship between
the maximum annual landings in a time-series for a given species and the MSY as
estimated through traditional stock assessment methods.}
\usage{
estimate_msy(max_catch)
}
\arguments{
\item{max_catch}{The maximum catch of a time-series}
}
\description{
Estimate MSY
A function to estimate Maximum Sustainable Yield from a time series of landings
The function requires only the maximum catch of a time-series.
Based on the work of Cheung & Srinivasan who find a statistical relationship between
the maximum annual landings in a time-series for a given species and the MSY as
estimated through traditional stock assessment methods.
}
\keyword{MSY,}
\keyword{Maximum}
\keyword{Sustainable}
\keyword{Yield}

