% Generated by roxygen2 (4.0.2): do not edit by hand
\name{landings_sma}
\alias{landings_sma}
\title{Landings SMA
This function is an adjusted SMA with the ends filled.
Avoids the case of missing data (NAs) at the beginning and end of the data series.}
\usage{
landings_sma(x, n = 5)
}
\arguments{
\item{x}{A timeseries of data}

\item{n}{The number of observations to average. Default is 5 period moving average.}
}
\description{
Landings SMA
This function is an adjusted SMA with the ends filled.
Avoids the case of missing data (NAs) at the beginning and end of the data series.
}
\examples{
sma(c(1:100))
}
\keyword{average}
\keyword{moving}
\keyword{simple}

