% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/of_style.R
\name{odsk_style}
\alias{odsk_style}
\title{Add odsk theme to ggplot chart}
\usage{
odsk_style()
}
\description{
This function allows you to add the odsk theme to your ggplot graphics.
Based on the {bbplot} package
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
odsk_style()
}
\keyword{odsk_style}
