% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenfisheries.R
\docType{package}
\name{ropenfisheries-package}
\alias{ropenfisheries}
\alias{ropenfisheries-package}
\title{ropenfisheries: What the package does (short line)}
\description{
An R package for global fisheries research and analysis.
}
\details{
The OpenFisheries R package.
}
\keyword{FAO}
\keyword{data}
\keyword{fisheries}
\keyword{openfisheries}
