% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sma}
\alias{sma}
\title{Simple Moving Average}
\usage{
sma(x, n = 5)
}
\arguments{
\item{x}{A timeseries of data}

\item{n}{The number of observations to average. Default is 5 period moving average.}
}
\description{
A function to compute a simple moving average
}
\examples{
sma(c(1:100))
}
\keyword{average}
\keyword{moving}
\keyword{simple}

