% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\name{stock_status}
\alias{stock_status}
\title{Stock Status
This function takes a landings time series from 1950 to 2011 as the argument 
Outputs the time-series of stock-status, MSY, and lost potential catch}
\usage{
stock_status(data, n = 5)
}
\arguments{
\item{data}{a time-series of landings.}
}
\description{
Stock Status
This function takes a landings time series from 1950 to 2011 as the argument 
Outputs the time-series of stock-status, MSY, and lost potential catch
}
